Add logging, update solution structure, and config changes
- Updated `MatLidStore.sln` to change project type GUIDs, add solution folders, and nest projects.
- Added `StatusCodes.Status404NotFound` response type to `HttpPost` methods in `AddressController.cs` and `UserController.cs`.
- Changed `[Authorize]` to `[AllowAnonymous]` in `MatLidStoreBaseController.cs`.
- Set `launchBrowser` to `true` in `launchSettings.json`.
- Updated `appsettings.Development.json` with new `EmailSettings`.
- Added logging functionality to various command and query handlers using `IAppLogger<T>`.
- Implemented logging for validation errors, not found exceptions, and successful operations.

+++ src/MSL.Application/Features/Comment/Commands/DeleteCommentCommand/DeleteCommentCommandHandler.cs
@@ -1,22 +1,34 @@
 ﻿using MediatR;
+using MLS.Application.Contracts.Logging;
 using MLS.Application.Contracts.Persistence.IRepositories;
+using MLS.Application.Exceptions;
 
 namespace MLS.Application.Features.Comment.Commands.DeleteCommentCommand;
 
 public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, Unit>
 {
     private readonly ICommentRepository _commentRepository;
+    private readonly IAppLogger<DeleteCommentCommandHandler> _logger;
 
-    public DeleteCommentCommandHandler(ICommentRepository commentRepository)
+    public DeleteCommentCommandHandler(ICommentRepository commentRepository, IAppLogger<DeleteCommentCommandHandler> logger)
     {
         _commentRepository = commentRepository;
+        _logger = logger;
     }
 
     public async Task<Unit> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
     {
         var commentToDelete = await _commentRepository.GetByIdAsync(request.Id);
+
+        if (commentToDelete is null)
+        {
+            _logger.LogWarning("Object {0} with id value equal to {1} was not found in the delete request.", nameof(Comment), request.Id);
+            throw new NotFoundException(nameof(Domain.Entities.Comment), request.Id);
+        }
+
         await _commentRepository.DeleteAsync(commentToDelete);
 
+        _logger.LogInformation("Comment was deleted successfully!");
         return Unit.Value;
     }
 }
